% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{A_t_inv_two}
\alias{A_t_inv_two}
\title{Compute inverse divergence function for a(t) = c / (1-t)^2}
\usage{
A_t_inv_two(c, y)
}
\arguments{
\item{c}{a positive number for the divergence hyperparameter. A larger value implies
earlier divergence on the tree}

\item{y}{a positive number to take inverse}
}
\value{
a number in the interval (0, 1)
}
\description{
Compute inverse divergence function for a(t) = c / (1-t)^2
}
\seealso{
Other divergence functions: 
\code{\link{A_t_inv_one}()},
\code{\link{a_t_one_cum}()},
\code{\link{a_t_one}()},
\code{\link{a_t_two_cum}()},
\code{\link{a_t_two}()}
}
\concept{divergence functions}
