% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{a_t_one}
\alias{a_t_one}
\title{Compute divergence function a(t) = c / (1-t)}
\usage{
a_t_one(c, t)
}
\arguments{
\item{c}{a positive number for the divergence hyperparameter. A larger value implies
earlier divergence on the tree}

\item{t}{a number in the interval (0, 1) indicating the divergence time}
}
\value{
a positive number
}
\description{
Compute divergence function a(t) = c / (1-t)
}
\seealso{
Other divergence functions: 
\code{\link{A_t_inv_one}()},
\code{\link{A_t_inv_two}()},
\code{\link{a_t_one_cum}()},
\code{\link{a_t_two_cum}()},
\code{\link{a_t_two}()}
}
\concept{divergence functions}
