% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_leaf_branch}
\alias{add_leaf_branch}
\title{add a leaf branch to an existing tree tree_old
div_t: divergence time of the new branch
new_leaf_label: the label of the newly added leaf
where: node name of to which node in the existing tree the new leaf branch should connect to
position: the numerical location of the left side of the added branch}
\usage{
add_leaf_branch(tree_old, div_t, new_leaf_label, where, position)
}
\description{
add a leaf branch to an existing tree tree_old
div_t: divergence time of the new branch
new_leaf_label: the label of the newly added leaf
where: node name of to which node in the existing tree the new leaf branch should connect to
position: the numerical location of the left side of the added branch
}
