% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_multichotomous_tip}
\alias{add_multichotomous_tip}
\title{add a leaf branch to an existing tree tree_old to make a multichotomus branch
div_t: divergence time of the new branch
new_leaf_label: the label of the newly added leaf
where: node name of to which node in the existing tree the new leaf branch should connect to}
\usage{
add_multichotomous_tip(tree_old, div_t, new_leaf_label, where)
}
\description{
add a leaf branch to an existing tree tree_old to make a multichotomus branch
div_t: divergence time of the new branch
new_leaf_label: the label of the newly added leaf
where: node name of to which node in the existing tree the new leaf branch should connect to
}
