% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_DDT_functions.R
\name{add_one_sample}
\alias{add_one_sample}
\title{Functions to simulate trees and node parameters from a DDT
process.
if t2 > t1, then select which path to take, with probability proportional to the number of data
points that already traversed the path}
\usage{
add_one_sample(tree_old, c, c_order, theta, alpha)
}
\description{
Functions to simulate trees and node parameters from a DDT
process.
if t2 > t1, then select which path to take, with probability proportional to the number of data
points that already traversed the path
}
