% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_tree.R
\name{attach_subtree}
\alias{attach_subtree}
\title{randomly attach a subtree to a given DDT
subtree: subtree to attach to tree_kept
tree_kept: the tree to be attached to
detach_div_time: divergence time of subtree when it was extracted from the original tree
pa_detach_node_label: label of the parent node of the detached node}
\usage{
attach_subtree(
  subtree,
  tree_kept,
  detach_div_time,
  pa_detach_node_label,
  c,
  c_order = 1,
  theta = 0,
  alpha = 0
)
}
\description{
randomly attach a subtree to a given DDT
subtree: subtree to attach to tree_kept
tree_kept: the tree to be attached to
detach_div_time: divergence time of subtree when it was extracted from the original tree
pa_detach_node_label: label of the parent node of the detached node
}
