% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_hchs}
\alias{data_hchs}
\title{Parameters for the HCHS dietary recall data example}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(data_hchs)
}
\description{
A list of five variables containing the food items and parameter estimates obtained from
MCMC chains. The real multivariate binary dataset is not included for privacy.
The variables are as follows:
}
\details{
\itemize{
\item item_membership_list. A list of G = 7 elements, where the g-th element contains the
indices of items belonging to major food group g.
\item item_name_list. A list of G = 7 elements, where the g-th element contains the item
labels of items in major food group g, and the name of the g-th element is the major food
group label.
\item tree_phylo. The maximum a posterior tree estimate with K = 6 leaves obtained from the
real HCHS data. Class "phylo".
\item class_probability. A K-vector with entries between 0 and 1. The posterior mean estimate
for class probabilities obtained from the real HCHS data.
\item Sigma_by_group. A G-vector greater than 0. The posterior mean estimate for group-specific
diffusion variances obtained from the real HCHS data.
}
}
\references{
\url{https://arxiv.org/abs/2306.04700}
}
\keyword{datasets}
