% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtlcm-package.R
\docType{package}
\name{ddtlcm-package}
\alias{ddtlcm}
\alias{ddtlcm-package}
\title{ddtlcm: Latent Class Analysis with Dirichlet Diffusion Tree Process Prior}
\description{
Implements a Bayesian algorithm for overcoming weak separation in Bayesian latent class analysis. Reference: Li et al. (2023) \href{https://arxiv.org/abs/2306.04700}{arXiv:2306.04700}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/limengbinggz/ddtlcm}
  \item Report bugs at \url{https://github.com/limengbinggz/ddtlcm/issues}
}

}
\author{
\strong{Maintainer}: Mengbing Li \email{mengbing@umich.edu} (\href{https://orcid.org/0000-0002-2264-8006}{ORCID})

Other contributors:
\itemize{
  \item Briana Stephenson \email{bstephenson@hsph.harvard.edu} [contributor]
  \item Zhenke Wu \email{zhenkewu@umich.edu} (\href{https://orcid.org/0000-0001-7582-669X}{ORCID}) [contributor]
}

}
\keyword{internal}
