% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_mnorm}
\alias{draw_mnorm}
\title{sample multivariate normal using precision matrix
from x ~ N(Q^{-1}a, Q^{-1}), where Q^{-1} is the precision matrix}
\usage{
draw_mnorm(precision_mat, precision_a_vec)
}
\description{
sample multivariate normal using precision matrix
from x ~ N(Q^{-1}a, Q^{-1}), where Q^{-1} is the precision matrix
}
