% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exp_normalize}
\alias{exp_normalize}
\title{compute normalized probabilities: exp(x_i) / sum_j exp(x_j)}
\usage{
exp_normalize(x)
}
\arguments{
\item{x}{a rea-valued vector}
}
\description{
compute normalized probabilities: exp(x_i) / sum_j exp(x_j)
}
