% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialize_poLCA}
\alias{initialize_poLCA}
\title{Estimate an initial response profile from latent class model using poLCA()}
\usage{
initialize_poLCA(K, data, ...)
}
\arguments{
\item{K}{number of latent classes}

\item{data}{a N by J observed binary matrix, where the i,j-th element is the response
of item j for individual i}

\item{...}{optional arguments for the poLCA function}
}
\value{
a K by J probability matrix, the k,j-th entry being the response probability to item
j of an individual in class k
}
\description{
Estimate an initial response profile from latent class model using poLCA()
}
\seealso{
Other initialization functions: 
\code{\link{initialize_hclust}()},
\code{\link{initialize}()}
}
\concept{initialization functions}
