% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialize_randomLCM}
\alias{initialize_randomLCM}
\title{Provide a random initial response profile based on latent class mode}
\usage{
initialize_randomLCM(K, data)
}
\arguments{
\item{K}{number of latent classes}

\item{data}{a N by J observed binary matrix, where the i,j-th element is the response
of item j for individual i}
}
\value{
a K by J probability matrix, the k,j-th entry being the response probability to item
j of an individual in class k
}
\description{
Provide a random initial response profile based on latent class mode
}
