% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehoods.R
\name{J_n}
\alias{J_n}
\title{compute factor in the exponent of the divergence time distribution
l = number of data points to the left
r = number of data points to the right}
\usage{
J_n(l, r)
}
\description{
compute factor in the exponent of the divergence time distribution
l = number of data points to the left
r = number of data points to the right
}
