% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_expit}
\alias{log_expit}
\title{Numerically accurately compute f(x) = log(x / (1/x)).}
\usage{
log_expit(x)
}
\arguments{
\item{x}{a value or a numeric vector between 0 and 1 (exclusive)}
}
\value{
a number or rea-valued vector
}
\description{
Numerically accurately compute f(x) = log(x / (1/x)).
}
