% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{The logistic function}
\usage{
logit(x)
}
\arguments{
\item{x}{a value or a numeric vector between 0 and 1 (exclusive)}
}
\description{
The logit function: f(x) = log(x / (1/x)). Large absolute values of
x will be truncated to +/- 5 after logit transformation according to its sign.
}
