% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehoods.R
\name{logllk_ddt}
\alias{logllk_ddt}
\title{compute the loglihood of the entire DDT}
\usage{
logllk_ddt(
  c,
  c_order,
  Sigma_by_group,
  tree_phylo4d,
  item_membership_list,
  tree_structure_old = NULL,
  dist_mat_old = NULL
)
}
\description{
compute the loglihood of the entire DDT
}
