% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehoods.R
\name{logllk_div_time_one}
\alias{logllk_div_time_one}
\title{compute loglikelihood of divergence times for a(t) = c/(1-t)}
\usage{
logllk_div_time_one(c, l, r, t)
}
\description{
compute loglikelihood of divergence times for a(t) = c/(1-t)
}
