% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehoods.R
\name{logllk_div_time_two}
\alias{logllk_div_time_two}
\title{compute loglikelihood of divergence times for a(t) = c/(1-t)^2}
\usage{
logllk_div_time_two(c, l, r, t)
}
\description{
compute loglikelihood of divergence times for a(t) = c/(1-t)^2
}
