% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehoods.R
\name{logllk_location}
\alias{logllk_location}
\title{compute the marginal log likelihood of the locations
tree_phylo4d: tree in the phylo4d class
Sigma_by_group: variance of the brownian motion
dist_mat_old: a known pairwise MRCA matrix between leaf nodes, to save computation time}
\usage{
logllk_location(
  tree_phylo4d,
  Sigma_by_group,
  item_membership_list,
  dist_mat = NULL,
  tol = 1e-07
)
}
\description{
compute the marginal log likelihood of the locations
tree_phylo4d: tree in the phylo4d class
Sigma_by_group: variance of the brownian motion
dist_mat_old: a known pairwise MRCA matrix between leaf nodes, to save computation time
}
