% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehoods.R
\name{logllk_tree_topology}
\alias{logllk_tree_topology}
\title{Compute loglikelihood of the tree topology}
\usage{
logllk_tree_topology(l, r)
}
\arguments{
\item{l}{number of data points to the left}

\item{r}{number of data points to the right}
}
\description{
Compute loglikelihood of the tree topology
}
\seealso{
Other likelihood functions: 
\code{\link{logllk_ddt_lcm}()},
\code{\link{logllk_ddt}()},
\code{\link{logllk_div_time_one}()},
\code{\link{logllk_div_time_two}()},
\code{\link{logllk_lcm}()},
\code{\link{logllk_location}()}
}
\concept{likelihood functions}
