% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot}
\alias{plot}
\title{Generic function}
\usage{
plot(x, log = TRUE, ...)
}
\arguments{
\item{x}{The object to print}

\item{log}{Default argument passed to plot(). Not used.}

\item{\dots}{Further arguments passed to each method}
}
\description{
Generic function
}
