% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ddt_lcm}
\alias{predict.ddt_lcm}
\title{Prediction of class memberships from posterior predictive distributions}
\usage{
\method{predict}{ddt_lcm}(object, data, burnin = 3000, ...)
}
\arguments{
\item{object}{a ddt_lcm object}

\item{data}{an NxJ matrix of multivariate binary responses, where
N is the number of individuals, and J is the number of granular items}

\item{burnin}{number of samples to discard from the posterior chain as burn-ins. Default is 3000.}

\item{\dots}{Further arguments passed to each method}
}
\description{
Predict individual class memberships based on posterior predictive distributions.
For each posterior sample, let the class memberships be modal assignments. Then aggregate over
all posterior samples to obtain the most likely assigned classes.
}
