% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Generic function}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{The object to print}

\item{\dots}{Further arguments passed to each method}
}
\description{
Generic function
}
