% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.summary.ddt_lcm}
\alias{predict.summary.ddt_lcm}
\title{Prediction of class memberships from posterior summaries}
\usage{
\method{predict}{summary.ddt_lcm}(object, data, ...)
}
\arguments{
\item{object}{a "summary.ddt_lcm" object}

\item{data}{an NxJ matrix of multivariate binary responses, where
N is the number of individuals, and J is the number of granular items}

\item{\dots}{Further arguments passed to each method}
}
\description{
Predict individual class memberships based on posterior summary (point estimates
of model parameters). The predicted class memberships are modal assignments.
}
\examples{
data(result_hchs)
burnin <- 50
summarized_result <- summary(result_hchs, burnin, relabel = TRUE, be_quiet = TRUE)
predicted <- predict(summarized_result, result_hchs$data)
}
