% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.summary.ddt_lcm_all}
\alias{predict.summary.ddt_lcm_all}
\title{Prediction of class memberships from posterior summaries}
\usage{
\method{predict}{summary.ddt_lcm_all}(model_summary, data)
}
\arguments{
\item{model_summary}{a "summary.ddt_lcm_all" object}

\item{data}{an NxJ matrix of multivariate binary responses, where
N is the number of individuals, and J is the number of granular items}
}
\description{
Predict individual class memberships based on posterior summary (point estimates
of model parameters). The predicted class memberships are modal assignments.
}
