% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{print.ddt_lcm}
\alias{print.ddt_lcm}
\title{Print out a ddt_lcm model}
\usage{
\method{print}{ddt_lcm}(model, digits = 3L)
}
\arguments{
\item{model}{a "ddt_lcm" object}
}
\description{
Print out a ddt_lcm model
}
\seealso{
Other ddt_lcm results: 
\code{\link{print.summary.ddt_lcm_all}()},
\code{\link{summary.ddt_lcm}()}
}
\concept{ddt_lcm results}
