% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print}
\alias{print}
\title{Generic function}
\usage{
print(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{\dots}{Further arguments passed to each method}
}
\description{
Generic function
}
