% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.summary.ddt_lcm}
\alias{print.summary.ddt_lcm}
\title{Print out summary of a ddt_lcm model}
\usage{
\method{print}{summary.ddt_lcm}(x, digits = 3L, ...)
}
\arguments{
\item{x}{a "summary.ddt_lcm" object}

\item{digits}{integer indicating the number of decimal places (round) to be used.}
}
\description{
Print out summary of a ddt_lcm model
}
\seealso{
Other ddt_lcm results: 
\code{\link{print.ddt_lcm}()},
\code{\link{summary.ddt_lcm}()}
}
\concept{ddt_lcm results}
