% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{print.summary.ddt_lcm_all}
\alias{print.summary.ddt_lcm_all}
\title{Print out summary of a ddt_lcm model}
\usage{
\method{print}{summary.ddt_lcm_all}(model_summary, digits = 3L)
}
\arguments{
\item{model_summary}{a "summary.ddt_lcm_all" object}
}
\description{
Print out summary of a ddt_lcm model
}
\seealso{
Other ddt_lcm results: 
\code{\link{print.ddt_lcm}()},
\code{\link{summary.ddt_lcm}()}
}
\concept{ddt_lcm results}
