% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress print from cat()}
\usage{
quiet(x, be_quiet = TRUE)
}
\arguments{
\item{x}{evaluation of a statement that may explicitly or implicitly involve cat()}

\item{be_quiet}{logical. TRUE to suppress print from cat(); FALSE to continue printing}
}
\description{
Suppress print from cat()
}
