% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_tree.R
\name{random_detach_subtree}
\alias{random_detach_subtree}
\title{Metropolis-Hasting algorithm for sampling tree topology and
branch lengths from the DDT branching process.
randomly detach a subtree from a given tree}
\usage{
random_detach_subtree(tree_phylo4)
}
\description{
Metropolis-Hasting algorithm for sampling tree topology and
branch lengths from the DDT branching process.
randomly detach a subtree from a given tree
}
