% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_tree.R
\name{reattach_point}
\alias{reattach_point}
\title{choose a divergence time, location, and tree topology to reattach a subtree
tree_kept: the tree to be attached to}
\usage{
reattach_point(tree_kept, c, c_order = 1, theta = 0, alpha = 0)
}
\description{
choose a divergence time, location, and tree topology to reattach a subtree
tree_kept: the tree to be attached to
}
