% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result_hchs}
\alias{result_hchs}
\title{Result of fitting DDT-LCM to a semi-synthetic data example}
\format{
A list with 8 elements
}
\usage{
data(result_hchs)
}
\description{
This is a "ddtlcm" object obtained from running \code{ddtlcm_fit} to a semi-synthetic dataset with
100 posterior samples (for the sake of time). See \code{\link{ddtlcm_fit}} for description of
the object.
}
\keyword{datasets}
