% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_tree.R
\name{sample_c_two}
\alias{sample_c_two}
\title{Sample divergence function parameter c for a(t) = c / (1-t)^2 through Gibbs sampler}
\usage{
sample_c_two(shape0, rate0, tree_structure)
}
\arguments{
\item{shape0}{shape of the Inverse-Gamma prior}

\item{rate0}{rate of the Inverse-Gamma prior}

\item{tree_structure}{a data.frame containing the divergence times and number of
data points to the left and right branches of internal nodes on the tree}
}
\value{
a numeric value of the newly sampled c
}
\description{
Sample divergence function parameter c for a(t) = c / (1-t)^2 through Gibbs sampler
}
