% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_lcm.R
\name{sample_class_assignment}
\alias{sample_class_assignment}
\title{Sample individual class assignments Z_i, i = 1, ..., N}
\usage{
sample_class_assignment(
  data,
  leaf_data,
  a_pg,
  auxiliary_mat,
  class_probability
)
}
\arguments{
\item{data}{a N by J binary matrix, where the i,j-th element is the response
of item j for individual i}

\item{leaf_data}{a K by J matrix of logit(theta_{kj})}

\item{a_pg}{a N by J matrix of hyperparameters of the generalized logistic distribution}

\item{auxiliary_mat}{a N by J matrix of truncated normal variables from previous iteration}

\item{class_probability}{a length K vector, where the k-th element is the
probability of assigning an individual to class k. It does not have to sum up to 1}
}
\value{
a vector of length N, where the i-th element is the class assignment of
individual i
}
\description{
Sample individual class assignments Z_i, i = 1, ..., N
}
