% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_lcm.R
\name{sample_sigmasq}
\alias{sample_sigmasq}
\title{Sample item group-specific variances through Gibbs sampler}
\usage{
sample_sigmasq(shape0, rate0, dist_mat, item_membership_list, locations)
}
\arguments{
\item{shape0}{a vector of G elements, each being the shape of the
Inverse-Gamma prior of group g}

\item{rate0}{a vector of G elements, each being the rate of the
Inverse-Gamma prior of group g}

\item{dist_mat}{a list, containing the KxK tree-structured matrix of leaf nodes,
where K is the number of leaves / latent classes, and SVD components}

\item{item_membership_list}{a vector of G elements, each indicating the number of
items in this group}

\item{locations}{a KxJ matrix of leaf parameters}
}
\value{
a numeric vector of G elements, each being the newly sampled variance
of the latent location of this group
}
\description{
Sample item group-specific variances through Gibbs sampler
}
