% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_DDT_functions.R
\name{simulate_parameter_on_tree}
\alias{simulate_parameter_on_tree}
\title{Simulate node parameters along a given tree.}
\usage{
simulate_parameter_on_tree(
  tree_phylo,
  Sigma_by_group,
  item_membership_list,
  root_node_location = 0
)
}
\arguments{
\item{tree_phylo}{a "phylo" object containing the tree topology and branch lengths}

\item{Sigma_by_group}{a G-vector greater than 0. The initial values for the group-specific diffusion
variances}

\item{item_membership_list}{a list of G elements, where the g-th element contains the indices of
items in major group g}

\item{root_node_location}{the coordinate of the root node parameter. By default, the node parameter
initiates at the origin so takes value 0. If a value, then the value will be repeated into a
length J vector. If a vector, it must be of length J.}
}
\value{
A class "phylo4d" tree with K leaves with node parameters. The leaf nodes are labeled "v1", ..., "vK",
root node "u1", and internal nodes "u2", ..., "uK".
}
\description{
Simulate node parameters along a given tree.
}
\seealso{
Other simulate DDT-LCM data: 
\code{\link{simulate_DDT_tree}()},
\code{\link{simulate_lcm_given_tree}()},
\code{\link{simulate_lcm_response}()}
}
\concept{simulate DDT-LCM data}
