% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.ddt_lcm}
\alias{summary.ddt_lcm}
\title{Summarize the output of a ddt_lcm model}
\usage{
\method{summary}{ddt_lcm}(result, burnin = 3000, relabel = T, be_quiet = F)
}
\arguments{
\item{result}{a "ddt_lcm" object}

\item{burnin}{number of samples to discard from the posterior chain as burn-ins. Default is 3000.}

\item{relabel}{If TRUE, perform post-hoc label switching using the Equivalence Classes
Representatives (ECR) method to solve non-identifiability issue in mixture models. If FALSE,
no label switching algorithm will be performed.}

\item{be_quiet}{If TRUE, do not print information during summarization. If FALSE, print label
switching information and model summary.}
}
\description{
Summarize the output of a ddt_lcm model
}
\seealso{
Other ddt_lcm results: 
\code{\link{print.ddt_lcm}()},
\code{\link{print.summary.ddt_lcm}()}
}
\concept{ddt_lcm results}
