% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary}
\alias{summary}
\title{Generic function}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{The object to summarize}

\item{\dots}{Further arguments passed to each method}
}
\description{
Generic function
}
