% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_tree.R
\name{attach_subtree}
\alias{attach_subtree}
\title{Attach a subtree to a given DDT at a randomly selected location}
\usage{
attach_subtree(
  subtree,
  tree_kept,
  detach_div_time,
  pa_detach_node_label,
  c,
  c_order = 1,
  theta = 0,
  alpha = 0
)
}
\arguments{
\item{subtree}{subtree to attach to tree_kept}

\item{tree_kept}{the tree to be attached to}

\item{detach_div_time}{divergence time of subtree when it was extracted from the original tree}

\item{pa_detach_node_label}{label of the parent node of the detached node}
}
\description{
Attach a subtree to a given DDT at a randomly selected location
}
