% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exp_normalize}
\alias{exp_normalize}
\title{Compute normalized probabilities: exp(x_i) / sum_j exp(x_j)}
\usage{
exp_normalize(x)
}
\arguments{
\item{x}{a number or rea-valued vector}
}
\value{
a number or rea-valued vector
}
\description{
Compute normalized probabilities: exp(x_i) / sum_j exp(x_j)
}
