% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehoods.R
\name{J_n}
\alias{J_n}
\title{Compute factor in the exponent of the divergence time distribution}
\usage{
J_n(l, r)
}
\arguments{
\item{l}{number of data points to the left}

\item{r}{number of data points to the right}
}
\description{
Compute factor in the exponent of the divergence time distribution
}
