% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.ddt_lcm}
\alias{print.ddt_lcm}
\title{Print out setup of a ddt_lcm model}
\usage{
\method{print}{ddt_lcm}(x, ...)
}
\arguments{
\item{x}{a "ddt_lcm" object}

\item{\dots}{Further arguments passed to each method}
}
\description{
Print out setup of a ddt_lcm model
}
\examples{
data(result_diet_1000iters)
print(result_diet_1000iters)
}
\seealso{
Other ddt_lcm results: 
\code{\link{print.summary.ddt_lcm}()},
\code{\link{summary.ddt_lcm}()}
}
\concept{ddt_lcm results}
